old-school optimizations
int->tinyint for scores and number of tests

+++ dbv_data/schema/history_tm_submissions_tests.sql
@@ -3,7 +3,7 @@ CREATE TABLE `history_tm_submissions_tests` (
   `id` bigint(20) NOT NULL,
   `idSubmission` bigint(20) NOT NULL DEFAULT '0',
   `idTest` bigint(20) NOT NULL DEFAULT '0',
-  `iScore` int(11) NOT NULL DEFAULT '0',
+  `iScore` tinyint(3) NOT NULL DEFAULT '0',
   `iTimeMs` int(11) NOT NULL DEFAULT '0',
   `iMemoryKb` int(11) NOT NULL DEFAULT '0',
   `iErrorCode` int(11) NOT NULL DEFAULT '0',
